Thumbnail menu fixes, custom labels, .destop file hotkey support
Remove some more code that is not needed
Fixed more thumbnail menu bugs that I recently caused
Fixed opening/closing Thumbnail menu on hover

New Features:
Added custom label setting (in context menu) which allows the user
to configure a particular application to show the application name
or window title as the button label, over ridding the settings
default.

Added support for .desktop files names in hotkey descriptions which
will cause the hotkey to open new windows for applications that are
not currently running even if they are not pinned to the window list.