New mouse action to restore the last focused window for an app
Added a new mouse action option that will restore the window that was
last in focus (rather then the window attached to the button that was
clicked). This allows you to click on any application window list button
to active the most recently used window for that appication. In this way
you don't need to open the thumbnail window or remember which button
matches the application window that you were last interacting with.