Change how drag&drop of pools works.
When pooling is enabled and you drag a window list entry to a new
location outside the bounds of the pool, it will still move the entire
pool of buttons as before, but now it will maintain the order of the
windows in the pool.  To reorder the window list entries in the pool
you must drag a button only within the bounds of the pool.

Also tweaked the configuration menu a bit.