Add a context menu item for adding hotkeys; shrink pinned buttons
1. Added to the button context menu an option that will add a new Hotkey
to the hotkey list for the app. It will bring up the applet config
dialog and select the "Hotkeys" tab where the user can then select the
new Hotkey entry and set the key sequence and enable the hotkey.

2. Stop reserving label space for the indicator characters when pinned
buttons have no running windows and pinned button are configured to show
no label. I am not 100% convinced I like this behaviour better, but it
does save some space for my configuration.