Yet more Hotkeys context menu content changes & hotkey tooltips
1. Show a context menu option for buttons associated with an application
where the hotkey description matches the applications desktop file name
and the hotkey is NOT configured to cycle windows. This allows the user
to configure which window of an application is associated with the
hotkey. When "cycling" is enabled for an desktop file hotkey there is
no dynamic window assignment that makes any sense so (as before) the
context menu options will not appear for those cases.

2. Added a 2nd tooltip line showing the hotkey sequence for buttons that
are reachable via a hotkey.