Fix to prevent changing the window order during grouping/ungrouping
When grouping or ungrouping a set of application windows the order
of the windows as seen in the thumbnail menu or in an application
pool on the window list might change. This fix will maintain the
window order unless changed by the user using drag and drop. This
was a regression from a previous change.