New Ctrl/Shift+Click mouse settings + new move to current monitor action
1. Added an Advanced config settings tab with a table that allows the
user to add mouse actions for the Ctrl/Shift + mouse button
combinations. This adds 10 addition click actions which can be assigned
to windowlist button clicks or thumbnail window clicks. Users with
programmable mice could assign one of these Key+MouseButton combinations
to a programmable button and therefore have more single click actions.

There seems to be a Cinnamon bug with setting lists where there are
three or more options in a specific column when there are 3 or more
columns with option sets. The end result here is that the table with
the new advanced mouse options does not properly display the settings.
But this does not effect the operation of the feature and you can always
check what the correct setting are by opening the row edit dialog.
Hopefully I can find a fix for this when I have time to look into it
further

2. Added a mouse action option that will move the selected window
to the current monitor when the window is not on the same monitor
as the windowlist panel monitor. In the case where the window is already
on the windowlist panel monitor it will move the window the the next
monitor. Perfect for people with a 2 monitor setup.