Setup Wizard GUI and automatic config backup
1. Added a basic setup wizard GUI that will appear when you add a new
applet instance to a panel. This GUI will guide you through the basic
setup of the applet to achieve one of the typical window-list or
panel-laucher type setups without the complexity of the normal
applet configurator.

*NOTE*: For users who have an existing instance of CassiaWindowList
installed. Upgrading to this new code-base will cause the new Setup
Wizard to appear on the 1st restart of Cinnamon after upgrading the
applet code. You can just close the window and it will not appear
again on future restarts.

2. Added a new automatic backup feature that will backup the applet
configuration whenever any config settings are modified. The feature
is enabled when you enter a backup name under the "advanced" tab of
the configure window. A backed up config can be restored via the new
Setup Wizard which is displayed when you add a new instance of
CassiaWindowList to a panel. In this way you can easily restore
your setup in case you accidentally remove the Windowlist applet
or something corrupts your configuration.