Fix ungrouping and label pooling + features
1. Fixed the "Ungroup" action when the window-list was configured as
"Grouped" so that it will actually un-group the application. Previously
this option would only work when the window-list was configured as
"pooled" or "Automatic".

2. Fix some cases where the "one label per application pool" option
was not behaving correctly.

3. When only two workspaces exist, replace the "Pin to other workspaces"
context menu option to just be a toggle option for pinning to the other
workspace. The "Pin to other workspaces" sub-menu now only appears when
there are 3 or more workspaces

4. Added an option for "Pooled"/"Automatic" configurations that allows
users to decide if they want to see thumbnails for all applications in
the pool or just the thumbnail for just the one selected window.
"Show thumbnails for all windows of an application pool"

5. Added the word "Application" to the  "Un/Group windows" context menu
item to be a bit more clear

6. Make sure the last button in a "pool" is the pinned button so that
the label does not appear for pinned applications when the user selected
no labels for pinned buttons.