Change "Grouped" config ungrouping behaviour + fixes
1. When ungrouping in a "Grouped" configuration, add the new buttons
before the existing button being ungrouped. Previously they would be
added at the end of the windowlist. The window order when grouped
will match the order on the windowlist after ungrouping

2. Don't show the "Add new hotkey for..." context menu item for windows
that are only backed by a window (i.e. no .desktop file for the app).
Adding a hotkey for this type of window would not behave as expected

3. Fix some more cases where the "one label per application pool"
option was not behaving correctly.