Fixes and new grouped left click option
1. Added a new left click action option for grouped window-list buttons
called "Restore or hold for thumbnail menu". This option will restore
the most recent window on a click, open the Thumbnail menu when the
button is held for 350ms without mouse movement or (as per prior
behaviour) start a button drag operation if the mouse it moved before
the 350ms time expires.

2. Fix focus tracking while grouping and ungrouping an application when
the focus is currently with a window in the group.

3. Fix cases where the animated close would cause some errors when
changing the window-list config to/from pooled or grouped.