Fix for sticky urgent state and improved "move titlebar on to screen"
1. Fix (I think) for the issue where sometimes the "urgent" status of a
pinned button will remain even after the window that caused the urgent
status has closed. I have seen this sometimes when closing Steam using
the system tray icon when the Steam window is not open.

2. Improved the "Move titblebar on to screen" function so that it works
for all windows (even if it does not have a traditional titlebar, i.e.
Chrome, Calc, etc.), and now it will insure that the top-left of the
window is visible on the screen.