Added a new feature called "Smart Numeric Hotkeys".
1. Added a option in the "Hotkeys" setting tab (enabled by default)
called "Smart numeric hotkeys" which allows you to switch to the first
9 windows of a specific application (as ordered on the window list).
Creating a Hotkey containing the "1" key it will automatically be extended
to 1-9. For example, using Alt+1 with a "description" of "firefox.desktop"
will register Alt+1 through Alt+9 and using those hot keys will switch to
one of the first 9 firefox windows based on the order of the windows
on the window-list.

2. As part of the work in the above new feature, I rewrote much of the
hotkey handling which I believe has fixed a few other issues around
removing or reordering hotkeys. But it is still the case that manually
mapped hotkeys (using the button context menu option "Assign window to
a hotkey") will likely need to be reassigned in many cases.