More fixes for hotkeys
1. Allow hotkeys that have a "Description" matching an application name
to be manually assigned to any window. Only hotkeys with a desktop file
name "Description" can't be manually assigned now. The window cycling
feature is now based on the application type of the assigned window
rather then the application matching the hotkey "Description"

2. Fixed hotkeys using the "Smart numeric" feature so that the alternate
hotkey is not ignored

3. Fixed the tooltips so that "Smart numeric" hotkeys are always
(I hope) listed correctly in the tooltip

4. Fixed the "this._window is undefined" exception

5. Fixed more issues with incorrect hotkey assignments after removing or
rearranging hotkeys in the applet configuration dialog