Pinned button tooltips and Smart Numeric Hotkeys tooltip text
1. The Hotkey assigned to a pinned button will now show up in the
tooltip text for a pinned button that has no running windows as long
as the "Open new window when the application is not running" option
is enabled.

2. The tooltip text for the "Smart Numeric Hotkeys" configuration
option now has a more accurate description of the feature. The
hotkey description being an application name does not dictate which
windows the hotkeys will activate, only the application of the
assigned window makes that determination. The hotkey description
only helps automatically assign a window to the hotkey entry but
if that assignment is manually changed then the application of
the hotkeys also changes. For a hotkey description using a desktop
file name, the same process applies but there is no user option
to manually assign the hotkey to a window so effectively the
description does determine what windows it applies to.