Add "pin to launcher" context menu item
1. If an existing "launcher" applet exists on a panel (ether a 2nd
instance of CassiaWindowList setup in launcher mode or an instance of
panel-launchers@cinnamon.org) a new context menu item called
"Pin to launcher" will appear in the content menus. Using this context
menu item will add the application of the selected window-list button to
the launcher applet.

2. If an applet instance of panel-launchers@cinnamon.org exists but the
CassiaWindowList acquired the "panellauncher" role, it will forward any
"add to panel" requests to the panel-launchers applet rather the pinning
to the CassiaWindowList. Previously the same thing was done when a
2nd instance of CassiaWindowList (setup in launcher mode) existed.
This change just extends the same behaviour to cases where a
panel-launchers instance exists.