Change the context menu item for unpinning from a launcher instance
Previously the context menu option to unpin a button from a panel for
both the window-list and launcher modes was the same switch style item,
but that didn't make much sense for launcher mode. So I changed it
from a toggle item to a normal item when in launcher mode and changed
the text to "Remove from..." rather then "Pin to...".