Add new launcher mode left click option to launch new windows
Added a launcher specific configuration option that allows you to launch
a new instance of an application when using the left mouse button rather
then activating an existing window. This makes the launcher mode able to
behave more like a traditional panel-laucher. This new option will open
a new window on a click or open the thumbnail menu if the button is held
for 350ms. The new options is called "New window or hold for thumbnail
menu"

Also updated the README.md to reflect the current use of this github
repo now that the applet is available on cinnamon spices.