New "All Buttons" hotkey + hotkey hints + fixes
1. Added a special case hotkey description called "All Buttons" which
will cause the modifier+1-9 hotkeys to be mapped to the first 9 windows
in the window-list. Using this with the <Super> modifier will allow you
to have Super+1-9 work the same way that the standard Grouped Window
List works. I added a default hotkey for this new "All Buttons"
description but that will only appear in new instances (existing
instances will have to add one manually). Simply add a new hotkey with
key sequence that uses your desired modifier key(s) plus the "1" key and
set the description to "All Buttons".

2. Added a option to show hotkey hints when using the "`" (aka grave)
key with the hotkey modifiers. For example, if you define a hotkey using
<Super>+C assigned to the Calculator window, when you hit the <Super>+`
key the window-list button for Calculator will popup a "C" bubble to
show that the <Super>+C would select this window. If you defined hotkeys
using some other modifier sequence like <Super><Alt> then using
<Super><Alt>+` would show hints for those hotkeys.

3. Fixed a bug that prevented you from grouping a set of applications
windows when one of the application window-list buttons is pinned.

4. Fixed a bug with tooltips sometimes not show the correct set of
applicable hotkeys.