Fix for hotkey help bubbles and new window animations
1. Fixed some issues with the hotkey help bubbles:
Clear the window count bubbles before showing the hotkey-help bubbles so
that only buttons with hotkeys will have any bubbles shown when the
hotkey-help hotkey is used.

Clear the window count bubbles even when there is no hotkey-help to show
so that the user is not confused by window count bubbles.

2. Fixed the new window animation to only effect the window list button
if the button is a grouping button and therefore the new window will be
added to the existing window-list button.

3. Update the "Advanced" config tab help text box to document the
cinnamon fix that is required in order for the "Ctrl/Shift + mouse
button actions" list to correctly display the two middle column
contents.
https://github.com/linuxmint/cinnamon/pull/11908