Added option to show window from all workspaces
Added the option "Include window from all workspace" under the general
tab. This option will cause windows from other workspaces to be shown
in the window-list. Clicking on a window from another workspace will
caused the current workspace to switch to that windows workspace.
Rearranging the windows on one workspace does not effect the window
arrangement on other workspaces.