Add icon saturation option (grayscale/vivid)
1. Added a new option (under the Advanced tab in config) that allows the
user to modify the colour saturation of icons. A saturation of 0% will
result in grayscale icons and a saturation of 200% will make the icons
more vivid. The custom saturation (i.e something other then 100%) can
be applied to all buttons or just some types of buttons (minimized
buttons, idle pinned buttons, buttons on other workspaces or buttons on
other monitors). This options is experimental since there are some cases
where icons are unaffected by this setting so your mileage may vary.

2. Since markup style tooltips are not supported in older versions of
Cinnamon, I disabled marked up tooltips when running on Cinnamon
versions less then 5.0.

3. When using the new option to include windows from other workspaces
the "Move to other workspaces" context menu would show options for an
incorrect set of workspaces, I fixed that and included the text
"(this workspace)" to the menu item for the current workspace.