Add option to change icon saturation for all but the focused window
1. Added a new option for when the custom icon saturation will apply.
The option will use the custom saturation for all windows buttons
exception the one that currently has the focus.

2. Fixed the option that only applies the custom colour saturation to
buttons for windows on other monitors. Previously it would only  work
when new windows were added directly to the other monitors, but now the
changes also apply when moving a window to another monitor.

3. Optimized the icon custom saturation option by not loading and
modifying the icon for each switch between the normal icon and the
modified icon. Now the icon is only re-loaded when there is a real need.