Fix for the "All Buttons" hotkey when using translations
The "All Buttons" default hotkey entry uses an English "All Buttons"
which can't be translated but the hard-coded "all buttons" string in
the code is translated and therefore the default "All Buttons" entry
would not work unless the description string was changed to a
translation of "All Buttons". This fix allows both English and a
translation to work.