No click focus changing, Thumbnail menu fade effect & fade effect option
Added new options to automatically active windows when leaving the panel/menu (no click focus changing). This allows the user to activate a window by simply moving to mouse over a window-list button or a Thumbnail menu item and then moving back towards the middle of the screen. The action can be cancelled by holding the Ctrl or Shift keys, or by leaving the window-list or Thumbnail menu by moving the pointer off either end of the window-list/menu. The config option to enable this feature for the window-list is in the General tab and the option for the Thumbnail menu is in the Thumbnail tab

Added a fade-in/out effect is applied when opening/closing the Thumbnail menu

Don't make a hover clone if the window already has the focus

Add an option (advanced config tab) allowing the user to adjust the fade effect duration which is used when opening/closing the Thumbnail menu and when showing a hover peek full size window.

Changed all the effect duration options to have a maximum of 2000ms (2sec) and to have a step change of 20ms for each click on the -/+ adjustment buttons. The old max of 10sec was unreasonably high and a step change of just 1ms was kinda useless.