Prevent the Thumbnail menu from grabbing the focus
Preventing the Thumbnail menu from grabbing the focus has two positive effects; 1. It prevents odd cursor key behaviour when the menu is open. 2. It allows the workspace switching hotkeys (Ctrl+Alt+CursorKeys) to work when the menu is open. It is not 100% clear to be why the workspace switching was prevented when the menu was open, but this change seems to fix it so I hope that's the end of this story.