When possible, keep existing thumbnail menu when mouse hover changes
Don't close the existing thumbnail menu and open a new one when the mouse hover moves between different buttons in the same application pool (as long as the menu does have an item for the currently hovered button's window). This removes the useless thumbnail menu close then open sequence when using the application pooling behaviour and the thumbnail menus are configured to show all pooled windows.