Add option to show an ellipsis character for grouped buttons
Added an option for window-list buttons with more than 1 window (application groups) to show a unicode vertical ellipsis character to denote that the button represents a group of windows. This is an option that can be useful when people make use of the new "Number label contents" option to change the number label to represent something other than the "Application group window count". With this ellipsis, it is easy to see which buttons have more then one window when the number label is used to show workspace or monitor indexes rather then window count.