Added workspace & monitor numbers on thumbnail menu items
If there is more then one workspace and the configuration is setup to show windows from other workspaces then the workspace number will be shown inside round brackets on the top-right of all thumbnail menu items. The same type of logic is used for monitors. If both a workspace and monitor number needs to be shown, than (workspace#/Monitor#) will be shown. Currently this behaviour is unconditionally enabled, if you hate this and want an option to disable it please add a comment in issue #13 or open a new issue and let me know.