Allow the user to DND a window-list button to a different monitor
Allow the user to drag-and-drop a window-list button to a different monitor. If the drop target did not accept the drop and the currentWindow is not currently on the same monitor as where the drop occurred, then move the currentWindow of the window-list button to the target monitor and activate it. Currently this only works for window-list buttons and not thumbnail menu items, but I do plan to working on thumbnail menu drag-and-drop which would then allow the same "drag to a new monitor" feature to be possible as well.