Allow the user to DND a window-list button to the current workspace
As with the previous update for dragging to a new monitor, this commit does the same for workspaces. If you allow windows from other workspaces to appear on your window-list this change will allow you to drag-and-drop a window-list button from the windowlist on the the screen causing the window to be moved to the current workspace and then activated.