Added a "Move window here" context menu option
This new option button context menu option will move the selected window to the current workspace and/or monitor. It only appears when the button's window is not already on the current monitor/workspace. Obviously this is only possible when the config is setup so that a window from other monitors/workspaces will appear on the windowlist. After moving to this monitor and/or workspace it will then activate the windows to obtain the focus.