Added an option to set the mouse panel action for the scroll wheel
Added an configuration mouse action option to control the action taken when using the mouse scroll wheel on a window-list panel button. The options include changing the windows minimized/maximized state, the windows workspace, the windows monitor and the tiling location (the tiling option will only work in Cinnamon 5.4 or better). If the Thumbnail menu is open this action will be disabled and the thumbnail zooming effect will occur if it has been enabled. The default setting for this new option to is change the window from Minimized/Restored/Maximized state, but you can can change the option to be "none" if you prefer no scroll while action to be taken as it was before this feature was added.