Enable the "Close others" and "Close all" for any set of app windows
When there is more than one window-list buttons for the same application, the context menu items "Close others" and "Close all" will now appear. This allow you to close all windows for an application exception the selected button (close others) or close all the application windows. It occurs to me that this is a bit dangerous, I'll think of a way to make it harder to accidentally close a bunch of app windows. Maybe a long click needed to use these?