Fix the Thumbnail menu item animation on vertical panels
The Thumbnail menu item animation for removing the item from the menu
was unconditionally shrinking the width of the item to zero, this fix
will make the animation shrink the height to zero when the applet is on
a vertical panel.

I also changed the animation time from 0.5 sec to 0.2 sec.