Add "First character of window title" for the number label
This new option will show the 1st character of the window title using
the "number label" (making the name a bit incorrect now since it not
limited to just numbers). This is useful in cases where button labels
are disabled and you have more than one window for an application. Using
this option can give you a hint for which appliation window you want to
activate.